pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tire-extraction"
version = "1.0.0"
description = "Tire sidewall information extraction using computer vision and LLM"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.50.0",
    "ultralytics>=8.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "pillow-heif>=0.13.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
eval = [
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.21.0",
    "pandas>=2.0.0",
    "tqdm>=4.60.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
]

[tool.setuptools.packages.find]
include = ["src*"]
